dogsite-backend/dogsms/wsgi.py
import app
import google.cloud.logging

# Instantiates a client
client = google.cloud.logging.Client()

# Retrieves a Cloud Logging handler based on the environment
# you're running in and integrates the handler with the
# Python logging module. By default this captures all logs
# at INFO level and higher
client.setup_logging()

if __name__ == "__main__":
    app.app.run(host="0.0.0.0",port="8080")